RAG_01/advanced_rag_docs/requirements.txt
langchain
langchain-community
langchain-mistralai
faiss-cpu
beautifulsoup4
python-dotenv
lxml
streamlit